requirements.txt
streamlit
pandas
yfinance
plotly.express
alpha-vantage
stocknews